+++ examples/tutorial_add_module/1_custom_post_processing.cpp
@@ -78,7 +78,7 @@ int openPoseTutorialWrapper4()
 
         // OpenPose wrapper
         op::log("Configuring OpenPose wrapper...", op::Priority::Low, __LINE__, __FUNCTION__, __FILE__);
-        op::Wrapper<std::vector<op::UserDatum>> opWrapper;
+        op::WrapperT<std::vector<op::UserDatum>> opWrapper;
         // Pose configuration (use WrapperStructPose{} for default and recommended configuration)
         const op::WrapperStructPose wrapperStructPose{
             !FLAGS_body_disable, netInputSize, outputSize, keypointScale, FLAGS_num_gpu, FLAGS_num_gpu_start,